+++ frontend/scrum/src/components/common/LoginModal.module.css
@@ -1,0 +1,3 @@
+.loginButton {
+  cursor: pointer;
+}
