+++ frontend/scrum/src/pages/HomePage/Home.module.css
@@ -1,0 +1,12 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+}
+
+.logo {
+  max-width: 100%;
+  height: auto;
+}
